put 'modified' check into ExhibitPage model
Possibly uncouth to use addErrors to track this kind of problem. It might not be really an error as a form validation thing, but it seems to do the broader job. Also, I'm not using the form element name so that it shows up more prettily in the flash message. Couth/Schmouth/Vermouth.